// src/app/models/grading/student-custom-fields-view.ts
import { PrintTemplateCustomFieldView } from './print-template-custom-field-view';

/**
 * BaseWebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * StudentCustomFieldsView
 */
export interface StudentCustomFieldsView {
    /**
     * Gets or Sets SchoolAcademicSessionStudentId
     */
    schoolAcademicSessionStudentId?: number;
    /**
     * Gets or Sets GradingTermId
     */
    gradingTermId?: number;
    /**
     * Gets or Sets Value
     */
    printTemplateCustomFieldView?: Array<PrintTemplateCustomFieldView>;
    
}