// src/app/models/grading/print-template-custom-field-view.ts
/**
 * BaseWebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * PrintTemplateCustomFieldView
 */
export interface PrintTemplateCustomFieldView { 
    /**
     * Gets or Sets PrintTemplateCustomFieldId
     */
    printTemplateCustomFieldId?: string;
    /**
     * Gets or Sets Value
     */
    value?: string;

    id?: string;
}